Separated the init modal and update modal into its own components
